+++ node/services/autocomplete.ts
@@ -1,6 +1,5 @@
 import type { Context } from '@vtex/api'
 
-import { compareApiResults } from '../utils/compareResults'
 import type { Clients } from '../clients'
 
 /**
@@ -58,24 +57,13 @@ export function fetchSearchSuggestions(ctx: Context<Clients>, query: string) {
     'Search Suggestions'
   )
 }
-
 export function fetchCorrection(ctx: Context<Clients>, query: string) {
   const { intelligentSearchApi, intsch } = ctx.clients
 
-  return compareApiResults(
-    () =>
-      intelligentSearchApi.fetchCorrection({
-        query,
-      }),
-    () =>
-      intsch.fetchCorrection({
-        query,
-      }),
-    ctx.vtex.production ? 10 : 100,
+  return withFallback(
+    () => intsch.fetchCorrection({ query }),
+    () => intelligentSearchApi.fetchCorrection({ query }),
     ctx.vtex.logger,
-    {
-      logPrefix: 'Correction',
-      args: { query },
-    }
+    'Correction'
   )
 }